LayoutTests/platform/ios-simulator/fast/dom/Document/xml-document-focus-expected.txt

Make sure the focus management API is available to XML documents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.__proto__ is not HTMLDocument.prototype
PASS document.__proto__ is Document.prototype
PASS document.hasFocus() is true
PASS document.activeElement is document.body
PASS document.activeElement is testElement
PASS document.hasFocus() is true
PASS successfullyParsed is true

TEST COMPLETE
